[#3] refactor : 파일 분리
util 및 hook으로 각각 볼드처리랑 이벤트 관련 로직 분리했습니다.
