+++ client/src/util/boldMaker.js
@@ -1,0 +1,7 @@
+export const boldMaker = (input, list) => {
+	return list.map((data) => {
+		let regex = new RegExp(input, "gim");
+		return data.sickNm.replace(regex, `<b>${input}<b>`)
+	})
+
+}
